% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_steps_interface.R
\name{All_Steps_Interface}
\alias{All_Steps_Interface}
\title{RNASeqEnvironmentSet}
\usage{
All_Steps_Interface(RNASeqRParam, which.trigger = "OUTSIDE",
  INSIDE.path.prefix = NA, RNASeqQualityAssessment.RUN = TRUE,
  RNASeqReadProcess.RUN = TRUE, RNASeqDifferentialAnalysis.RUN = TRUE,
  RNASeqGoKegg.RUN = TRUE, OrgDb.species, go.level = 3, input.TYPE.ID,
  KEGG.organism, check.s4.print = TRUE)
}
\arguments{
\item{RNASeqRParam}{S4 object instance of experiment-related
parameters}

\item{which.trigger}{Default value is \code{OUTSIDE}. User should not change
this value.}

\item{INSIDE.path.prefix}{Default value is \code{NA}. User should not change
this value.}

\item{RNASeqQualityAssessment.RUN}{Default value is \code{TRUE}. Set \code{FALSE}
to skip "Quality Assessment" step.}

\item{RNASeqReadProcess.RUN}{Default value is \code{TRUE}. Set \code{FALSE}
to skip "RNASeq Read Process" step.}

\item{RNASeqDifferentialAnalysis.RUN}{Default value is \code{TRUE}. Set \code{FALSE}
to skip "RNASeq Differential Analysis" step.}

\item{RNASeqGoKegg.RUN}{Default value is \code{TRUE}. Set \code{FALSE}
to skip "RNASeq Go & Kegg" step.}

\item{OrgDb.species}{the genome wide annotation packages of species on
Bioconductor. Currently, there are 19 supported genome wide annotation
packages of species.}

\item{go.level}{the depth of acyclic graph in GO analysis}

\item{input.TYPE.ID}{The gene name type in OrgDb.species annotation packahge.}

\item{KEGG.organism}{the species that are supported for KEGG analysis.
Currently, there are more than 5000 supported species genome.
Check the valid species terms on
https://www.genome.jp/kegg/catalog/org_list.html}

\item{check.s4.print}{Default \code{TRUE}. If \code{TRUE},
the result of checking \code{RNASeqRParam} will be reported in
'Rscript_out/Environment_Set.Rout'. If \code{FALSE}, the result of checking
\code{RNASeqRParam} will not be in
'Rscript_out/Environment_Set.Rout'.}
}
\value{
None
}
\description{
Set up the environment for the following RNA-Seq workflow in R shell\cr
  This function do 4 things :\cr
  \enumerate{
    \item Create file directories.\cr
    \item Install necessary tools. \cr
    \item Export 'RNASeq_bin/' to the R environment. \cr
    \item Check command of tools. \cr
  }
  First it will create 'gene_data/', 'RNASeq_bin/', 'RNASeq_results/',
  'Rscript/', 'Rscript_out/' directories. \cr Afterwards, 'Hisat2',
  'Stringtie', 'Gffcompare' will be installed under
  'RNASeq_bin/Download/' and be unpacked under 'RNASeq_bin/Unpacked/'. \cr
  'RNASeq_bin/' will be added to the R environment and
  validity of tools will be checked.\cr
  Any ERROR occurs will be reported and the program will be terminated.\cr
  If you want to set up the environment for the following RNA-Seq workflow
  in background, please see \code{RNASeqEnvironmentSet_CMD()} function.
}
\examples{
data(yeast)
\dontrun{
RNASeqEnvironmentSet(RNASeqRParam = yeast)}
}
\author{
Kuan-Hao Chao
}
