% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_steps_interface.R
\name{All_Steps_Interface_CMD}
\alias{All_Steps_Interface_CMD}
\title{All_Steps_Interface_CMD}
\usage{
All_Steps_Interface_CMD(RNASeqRParam, RNASeqQualityAssessment.RUN = TRUE,
  RNASeqReadProcess.RUN = TRUE, RNASeqDifferentialAnalysis.RUN = TRUE,
  RNASeqGoKegg.RUN = TRUE, OrgDb.species, go.level = 3, input.TYPE.ID,
  KEGG.organism, run = TRUE, check.s4.print = TRUE)
}
\arguments{
\item{RNASeqRParam}{S4 object instance of experiment-related
parameters}

\item{RNASeqQualityAssessment.RUN}{Default value is \code{TRUE}. Set \code{FALSE}
to skip "Quality Assessment" step.}

\item{RNASeqReadProcess.RUN}{Default value is \code{TRUE}. Set \code{FALSE}
to skip "RNASeq Read Process" step.}

\item{RNASeqDifferentialAnalysis.RUN}{Default value is \code{TRUE}. Set \code{FALSE}
to skip "RNASeq Differential Analysis" step.}

\item{RNASeqGoKegg.RUN}{Default value is \code{TRUE}. Set \code{FALSE}
to skip "RNASeq Go & Kegg" step.}

\item{run}{Default value is \code{TRUE}. If \code{TRUE},
'Rscript/Environment_Set.R' will be created and executed.
The output log will be stored in 'Rscript_out/Environment_Set.Rout'.
If \code{False}, 'Rscript/Environment_Set.R' will be
created without executed.}

\item{check.s4.print}{Default \code{TRUE}. If \code{TRUE},
the result of checking \code{RNASeqRParam} will be reported in
'Rscript_out/Environment_Set.Rout'. If \code{FALSE}, the result of checking
\code{RNASeqRParam} will not be in
'Rscript_out/Environment_Set.Rout'.}
}
\value{
None
}
\description{
A functios to run all the steps with in one function. This function execute
  in the background:\cr
  \enumerate{
    \item Create file directories.\cr
    \item Install necessary tools. \cr
    \item Export 'RNASeq_bin/' to the R environment. \cr
    \item Check command of tools. \cr
  }
  First it will create 'gene_data/', 'RNASeq_bin/', 'RNASeq_results/',
  'Rscript/', 'Rscript_out/' directories. \cr Afterwards, 'Hisat2',
  'Stringtie', 'Gffcompare' will be installed under
  'RNASeq_bin/Download/' and be unpacked under 'RNASeq_bin/Unpacked/'. \cr
  'RNASeq_bin/' will be added to the R environment and
  validity of tools will be checked.\cr
  Any ERROR occurs will be reported and the program will be terminated.\cr
  If you want to set up the environment for the following RNA-Seq workflow
  in R shell, please see \code{RNASeqEnvironmentSet()} function.
}
\examples{
data(yeast)
\dontrun{
RNASeqEnvironmentSet_CMD(yeast)}
}
\author{
Kuan-Hao Chao
}
