% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ballgown_analysis.R
\name{CheckBallgownObject}
\alias{CheckBallgownObject}
\title{Check ballgown object}
\usage{
CheckBallgownObject(path.prefix)
}
\arguments{
\item{path.prefix}{path prefix of 'gene_data/', 'RNAseq_bin/', 'RNAseq_results/', 'Rscript/' and 'Rscript_out/' directories}
}
\value{
None
}
\description{
Check ballgown object
}
