% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_installtool_rnaseqpipline.R
\name{CheckDirAll}
\alias{CheckDirAll}
\title{Check sample gene and binary directory}
\usage{
CheckDirAll(path.prefix, print = TRUE)
}
\description{
Check sample gene and binary directory
}
