% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CheckExperimentGroup}
\alias{CheckExperimentGroup}
\title{inner function}
\usage{
CheckExperimentGroup(input.path.prefix, independent.variable, control.group,
  experiment.group)
}
\description{
inner function
}
