% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CheckExperimentType}
\alias{CheckExperimentType}
\title{inner function : check experiment type}
\usage{
CheckExperimentType(experiment.type = NA_character_, print = TRUE)
}
\description{
inner function : check experiment type
}
