% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CheckIndependentVariable}
\alias{CheckIndependentVariable}
\title{inner function : check experiment type}
\usage{
CheckIndependentVariable(independent.variable = NA_character_,
  input.path.prefix = NA_character_, print = TRUE)
}
\description{
inner function : check experiment type
}
