% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CheckInputDirFiles}
\alias{CheckInputDirFiles}
\title{inner function : check input.path.prefix}
\usage{
CheckInputDirFiles(input.path.prefix = NA_character_,
  gene.name = NA_character_, sample.pattern = NA_character_)
}
\description{
inner function : check input.path.prefix
}
