% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CheckInputParam}
\alias{CheckInputParam}
\title{inner function : check whether input values are NA}
\usage{
CheckInputParam(path.prefix = NA, input.path.prefix = NA, gene.name = NA,
  sample.pattern = NA, experiment.type = NA, main.variable = NA,
  additional.variable = NA)
}
\description{
inner function : check whether input values are NA
}
