% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CheckInputPrefixPath}
\alias{CheckInputPrefixPath}
\title{inner function : check prefix.path}
\usage{
CheckInputPrefixPath(input.path.prefix = NA_character_)
}
\description{
inner function : check prefix.path
}
