% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CheckMainVar}
\alias{CheckMainVar}
\title{inner function : check}
\usage{
CheckMainVar(input.path.prefix = NA_character_,
  main.variable = NA_character_, experiment.type = NA_character_,
  treatment.num = NA_character_, tissue.num = NA_character_,
  cell_type.num = NA_character_, genotype.num = NA_character_,
  time.num = NA_character_, dosage.time = NA_character_, print = TRUE)
}
\description{
inner function : check
}
