% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{CheckPhenodata}
\alias{CheckPhenodata}
\title{inside function : check valifity of phenodata
must have column : "ids" + "2 column"
other column : "treatment", "tissue", "cell_type", "genotype", "time", "dosage"}
\usage{
CheckPhenodata(input.path.prefix = NA_character_, gene.name = NA_character_,
  sample.pattern = NA_character_, print = TRUE)
}
\description{
inside function : check valifity of phenodata
must have column : "ids" + "2 column"
other column : "treatment", "tissue", "cell_type", "genotype", "time", "dosage"
}
