% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_environment_set.R
\name{CheckToolAll}
\alias{CheckToolAll}
\title{CheckToolAll}
\usage{
CheckToolAll(path.prefix, print = TRUE)
}
\arguments{
\item{path.prefix}{path prefix of 'gene_data/', 'RNASeq_bin/',
'RNASeq_results/', 'Rscript/' and 'Rscript_out/' directories.}

\item{print}{If \code{TRUE}, detailed information will be printed.
If \code{FALSE}, detailed information will not be printed.}
}
\value{
None
}
\description{
Check whether 'Hisat2', 'Stringtie', 'Samtools' and 'Gffcompare'
  are installed on the workstation
}
\examples{
data(yeast)
CheckToolAll(yeast@path.prefix,
             print=TRUE)
}
