% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_installtool_rnaseqpipline.R
\name{CopyInputDir}
\alias{CopyInputDir}
\title{inner function : Copy input files directory}
\usage{
CopyInputDir(path.prefix, input.path.prefix, gene.name, sample.pattern,
  indexes.optional = indexes.optional)
}
\description{
inner function : Copy input files directory
}
