% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat2_samtools_stringtie_ballgown_related.R
\name{CreateHisat2Index}
\alias{CreateHisat2Index}
\title{Creating Hisat2 index
Creat Hisat2 index for further use}
\usage{
CreateHisat2Index(path.prefix, gene.name, sample.pattern,
  splice.site.info = TRUE, exon.info = TRUE)
}
\description{
Creating Hisat2 index
Creat Hisat2 index for further use
}
