% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ballgown_analysis.R,
%   R/raw_reads_visualization.R
\name{DEGCorrelationPlot}
\alias{DEGCorrelationPlot}
\alias{DEGCorrelationPlot}
\title{Plot correlation plo}
\usage{
DEGCorrelationPlot()

DEGCorrelationPlot()
}
\description{
Plot correlation plo

Plot correlation plo
}
