% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ballgown_Differential_Express_analysis.R
\name{DEGOAnalysis}
\alias{DEGOAnalysis}
\title{This package will autamatically install org.Hs.eg.db, org.Rn.eg.db, org.Mm.eg.db. If you want to use different OrgDb annotation species, please install that annotation package and attach to your session.}
\usage{
DEGOAnalysis(path.prefix, OrgDb.species)
}
\description{
This package will autamatically install org.Hs.eg.db, org.Rn.eg.db, org.Mm.eg.db. If you want to use different OrgDb annotation species, please install that annotation package and attach to your session.
}
