% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqRParam_show_method.R
\docType{methods}
\name{show}
\alias{show}
\title{Show the details about RNASeqRParam S4 obejct}
\usage{
\S4method{show}{RNASeqRParam}(object)
}
\arguments{
\item{x}{a RNASeqRParam object}
}
\value{
None
}
\description{
Show the details about RNASeqRParam S4 obejct
}
\examples{
data(yeast)
show(yeast)
}
