% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExperimentTypeChecking.R
\name{ExperimentTypeChecking}
\alias{ExperimentTypeChecking}
\title{Check the experiment type and seperate into three function}
\usage{
ExperimentTypeChecking(path.prefix, gene.name, sample.pattern, experiment.type,
  main.variable, additional.variable)
}
\description{
Check the experiment type and seperate into three function
}
