% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat2_samtools_stringtie_ballgown_related.R
\name{GffcompareRefSample}
\alias{GffcompareRefSample}
\title{Examine how the transcripts compare with the reference annotation}
\usage{
GffcompareRefSample(path.prefix, gene.name, sample.pattern)
}
\description{
Examine how the transcripts compare with the reference annotation
}
