% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat2_samtools_stringtie_ballgown_related.R
\name{Hisat2ReportAssemble}
\alias{Hisat2ReportAssemble}
\title{Report Hisat2 assemble rate}
\usage{
Hisat2ReportAssemble(path.prefix, gene.name, sample.pattern)
}
\description{
Report Hisat2 assemble rate
}
