% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_install_tools.R
\name{InstallAll}
\alias{InstallAll}
\title{Install Hisat2, StringTie, Gffcompare, Samtools}
\usage{
InstallAll(path.prefix, os.type)
}
\description{
Install Hisat2, StringTie, Gffcompare, Samtools
}
