% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_environment_set.R
\name{InstallGffcompareBinary}
\alias{InstallGffcompareBinary}
\title{Install Gffcompare binary}
\usage{
InstallGffcompareBinary(path.prefix, os.type)
}
\description{
Install Gffcompare binary
}
