% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_environment_set.R
\name{InstallStringTieBinary}
\alias{InstallStringTieBinary}
\title{Install stringtie binary}
\usage{
InstallStringTieBinary(path.prefix, os.type)
}
\description{
Install stringtie binary
}
