% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/rnaseq_raw_read_process_hisat2_samtools_stringtie_ballgown_related.R
\name{PreDECountTable}
\alias{PreDECountTable}
\title{converting stringtie ballogwn preprocessed data to count table}
\usage{
PreDECountTable(path.prefix, sample.pattern, python.variable.answer,
  python.variable.version, print = TRUE)
}
\description{
converting stringtie ballogwn preprocessed data to count table
}
