% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_installtool_rnaseqpipline.R
\name{ProgressGenesFiles}
\alias{ProgressGenesFiles}
\title{check 'gene_data' and subdirectory files exit}
\usage{
ProgressGenesFiles(path.prefix, gene.name, sample.pattern, print = TRUE)
}
\description{
check 'gene_data' and subdirectory files exit
}
