% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_ballgown_DESeq2_edgeR.R
\name{RNAseqBallgownDESeq2EdgeRProcess_CMD}
\alias{RNAseqBallgownDESeq2EdgeRProcess_CMD}
\title{Ballgown analysis for RNA-Seq workflow in background}
\usage{
RNAseqBallgownDESeq2EdgeRProcess_CMD(RNASeqWorkFlowParam,
  ballgown.log2FC = 1, ballgown.qval = 0.05, DESeq2.padj = 0.1,
  DESeq2.log2FC = 1, edgeR.pval = 0.05, edgeR.log2FC = 1,
  run = TRUE, check.s4.print = TRUE)
}
\arguments{
\item{RNASeqWorkFlowParam}{S4 object instance of experiment-related parameters}

\item{ballgown.log2FC}{Default \code{1}. Set the threshold of log2 fold change to filter out differential expressed gene.}

\item{ballgown.qval}{Default \code{0.05}. Set the threshold of q-value to filter out differential expressed gene.}

\item{run}{Default value is \code{TRUE}. If \code{TRUE}, 'Rscript/Environment_Set.R' will be created and executed. The output log will be stored in 'Rscript_out/Environment_Set.Rout'.
If \code{False}, 'Rscript/Environment_Set.R' will be created without executed.}

\item{check.s4.print}{Default \code{TRUE}. If \code{TRUE}, the result of checking \code{RNASeqWorkFlowParam} will be reported in 'Rscript_out/Environment_Set.Rout'. If \code{FALSE}, the result of checking \code{RNASeqWorkFlowParam} will not be in 'Rscript_out/Environment_Set.Rout'}
}
\value{
None
}
\description{
Use ballgown R package for statistical analysis of assembled transcriptomes, including flexible differential expression analysis, and sample FPKM visualization as well as pre differential analysis visualization for the following RNA-Seq workflow in background.
This function do 3 things :
1. Create ballgown object, and write "ballgown_FPKM_result.csv" file.
2. Find the differential expressed genes, and write "ballgown_FPKM_DE_result.csv"
3. Sample data visualization.
   For transcript-related : Creating transcript_count_per_gene_plot, transcript_length_plot.
   For sample profile : Creating Box plot, Violin plot, Frequency plot, PCA related plots, Correlation plots.
   For pre differential expressed gene analysis : Creating MA plot, Volcano plot.
   For differential express gene analysis : Creating Heatmap plot, PCA related plots.
If you want to run ballgown analysis for the following RNA-Seq workflow in R shell, please see \code{RNAseqBallgownDESeq2EdgeRProcess()} function.
}
\examples{
\dontrun{
input_file_dir <- system.file(package = "RNASeqWorkflow", "exdata")
exp <- RNASeqWorkFlowParam(path.prefix = "/tmp/", input.path.prefix = input_file_dir, genome.name = "hg19", sample.pattern = "SRR[0-9]",
                           experiment.type = "two.group", main.variable = "treatment", additional.variable = "cell")
RNAseqEnvironmentSet_CMD(RNASeqWorkFlowParam = exp)}
}
