% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_differential_analysis.R
\name{RNASeqDifferentialAnalysis}
\alias{RNASeqDifferentialAnalysis}
\title{RNASeqDifferentialAnalysis}
\usage{
RNASeqDifferentialAnalysis(path.prefix, genome.name, sample.pattern,
  independent.variable, case.group, control.group, ballgown.pval = 0.05,
  ballgown.log2FC = 1, TPM.pval = 0.05, TPM.log2FC = 1,
  DESeq2.pval = 0.1, DESeq2.log2FC = 1, edgeR.pval = 0.05,
  edgeR.log2FC = 1)
}
\arguments{
\item{path.prefix}{path prefix of 'gene_data/', 'RNASeq_bin/',
'RNASeq_results/', 'Rscript/' and 'Rscript_out/' directories}

\item{genome.name}{genome.name Variable of genome name defined in
this RNA-Seq workflow (ex. \code{genome.name}.fa, \code{genome.name}.gtf)}

\item{sample.pattern}{sample.pattern  Regular expression of
paired-end fastq.gz files under 'input_files/raw_fastq.gz'.
Expression not includes \code{_[1,2].fastq.gz}.}

\item{independent.variable}{independent variable for the biological
experiment design of two-group RNA-Seq workflow}

\item{case.group}{group name of the case group}

\item{control.group}{group name of the control group}

\item{ballgown.pval}{Default \code{0.05}. Set the threshold of ballgown
p-value to filter out differential expressed gene.}

\item{ballgown.log2FC}{Default \code{1}. Set the threshold of ballgown
log2 fold change to filter out differential expressed gene.}

\item{TPM.pval}{Default \code{0.05}. Set the threshold of TPM & student
t-test p-value to filter out differential expressed gene.}

\item{TPM.log2FC}{Default \code{1}. Set the threshold of TPM & student
t-test log2 fold change to filter out differential expressed gene.}

\item{DESeq2.pval}{Default \code{0.05}. Set the threshold of DESeq2
p-value to filter out differential expressed gene.}

\item{DESeq2.log2FC}{Default \code{1}. Set the threshold of DESeq2
log2 fold change to filter out differential expressed gene.}

\item{edgeR.pval}{Default \code{0.05}. Set the threshold of edgeR
p-value to filter out differential expressed gene.}

\item{edgeR.log2FC}{Default \code{1}. Set the threshold of edgeR log2
fold change to filter out differential expressed gene.}
}
\value{
None
}
\description{
This function will run differential analysis on ballgown,
  DESeq2 and edgeR in background. \cr
  This function do following things : \cr
  \enumerate{
    \item ballgown analysis \cr
    Raw reads are normalized into FPKM values \cr
    The main statistic test in ballgown is paramatic F-test comparing nested
    linear models \cr
    \item TPM & Student's t-test analysis \cr
    TPM normalization is calculated from the FPKM values in ballgown. \cr
    Independent sample t-test is used with TPM values as input. P-values will
     be produced after t-test. \cr
    Fold change values are calculated by the formula :  \cr
    "`average TPM experiment group values divide` /
     `average TPM control group values`". \cr
    \item DESeq2 analysis  \cr
     Median of rations normalization(MRN) is used in DESeq2 for raw reads
      count normalization.  \cr
     Sequencing depth and RNA composition is taken into consideration is this
      normalization method.  \cr
     The main statistic test in DESeq2 is negative binomial distribution. \cr
    \item edgeR analysis  \cr
     Raw reads are normalized by TMM and library size.
      (run \code{calcNormFactors()} to get a DGEList,
     and then run \code{cpm()} on that DGEList)  \cr
     The main statistic test in edgeR is trimmed mean of M-values(TMM).\cr
  }
  If you want to run differential analysis on ballgown, TPM normalization,
  DESeq2, edgeR for the following RNA-Seq workflow in background,
  please see \code{RNASeqDifferentialAnalysis()} function.
}
\examples{
data(yeast)
\dontrun{
RNASeqDifferentialAnalysis(path.prefix          = yeast@path.prefix,
                           genome.name          = yeast@genome.name,
                           sample.pattern       = yeast@sample.pattern,
                           independent.variable = yeast@independent.variable,
                           case.group           = yeast@case.group,
                           control.group        = yeast@control.group)}
}
\author{
Kuan-Hao Chao
}
