% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_GO_KEGG.R
\name{RNASeqGoKegg}
\alias{RNASeqGoKegg}
\title{GO functional analysis and KEGG pathway analysis for RNA-Seq workflow in R shell}
\usage{
RNASeqGoKegg(path.prefix, independent.variable, OrgDb.species,
  KEGG.organism)
}
\arguments{
\item{path.prefix}{path prefix of 'gene_data/', 'RNASeq_bin/', 'RNASeq_results/', 'Rscript/' and 'Rscript_out/' directories}

\item{independent.variable}{independent variable for the biological experiment design of two-group RNA-Seq workflow}

\item{OrgDb.species}{WIP}

\item{KEGG.organism}{WIP}
}
\value{
None
}
\description{
Set up the environment for the following RNA-Seq workflow in background.
This function do 4 things :
1. Create file directories.
2. Install necessary tools.
3. Export 'RNASeq_bin/' to the R environment.
4. Check command of tools.
First it will create 'gene_data/', 'RNASeq_bin/', 'RNASeq_results/', 'Rscript/', 'Rscript_out/' directories. Afterwards, 'Hisat2', 'Stringtie', 'Samtools',
'Gffcompare' will be installed under 'RNASeq_bin/'. 'RNASeq_bin/' will be added to the R environment and validity of tools will be checked. Any ERROR occurs will be reported and the program will be terminated.
If you want to set up the environment for the following RNA-Seq workflow in R shell, please see \code{RNASeqEnvironmentSet()} function.
}
\examples{
\dontrun{
input_file_dir <- system.file(package = "RNASeqWorkflow", "exdata")
exp <- RNASeqWorkFlowParam(path.prefix = "/tmp/", input.path.prefix = input_file_dir, genome.name = "hg19", sample.pattern = "SRR[0-9]",
                           experiment.type = "two.group", main.variable = "treatment", additional.variable = "cell")
RNASeqEnvironmentSet_CMD(RNASeqWorkFlowParam = exp)}
}
\author{
Kuan-Hao Chao
}
