% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_GO_KEGG.R
\name{RNASeqGoKegg_CMD}
\alias{RNASeqGoKegg_CMD}
\title{RNASeqGoKegg_CMD}
\usage{
RNASeqGoKegg_CMD(RNASeqRParam, OrgDb.species, go.level = 3,
  input.TYPE.ID, KEGG.organism, run = TRUE, check.s4.print = TRUE)
}
\arguments{
\item{RNASeqRParam}{S4 object instance of experiment-related
parameters}

\item{OrgDb.species}{the genome wide annotation packages of species on
Bioconductor. Currently, there are 19 supported genome wide annotation
packages of species.}

\item{go.level}{the depth of acyclic graph in GO analysis}

\item{input.TYPE.ID}{The gene name type in OrgDb.species annotation packahge.}

\item{KEGG.organism}{the species that are supported for KEGG analysis.
Currently, there are more than 5000 supported species genome.
Check the valid species terms on
https://www.genome.jp/kegg/catalog/org_list.html}

\item{run}{Default value is \code{TRUE}. If \code{TRUE},
'Rscript/Environment_Set.R' will be created and executed.
The output log will be stored in 'Rscript_out/Environment_Set.Rout'.
If \code{False}, 'Rscript/Environment_Set.R' will be created
without executed.}

\item{check.s4.print}{Default \code{TRUE}. If \code{TRUE},
the result of checking \code{RNASeqRParam} will be reported in
'Rscript_out/Environment_Set.Rout'. If \code{FALSE}, the result of checking
\code{RNASeqRParam} will not be in
'Rscript_out/Environment_Set.Rout'}
}
\value{
None
}
\description{
Run Gene Ontology(GO) and Kyoto Encyclopedia of Genes and Genomes(KEGG)
  analysis in background. \cr
  This function do Gene Ontology(GO) and Kyoto Encyclopedia of Genes and
  Genomes(KEGG) analysis : \cr
  \enumerate{
    \item Gene Ontology(GO) :\cr
     \enumerate{
       \item Do GO function classification analysis. \cr
       \item Do GO function enrichment analysis. \cr
       \item Visualization : bar plot, dot plot etc. \cr
     }
    \item Kyoto Encyclopedia of Genes and Genomes(KEGG) :\cr
     \enumerate{
       \item Do KEGG pathway enrichment analysis \cr
       \item Pathway visulization with \code{pathview} package. KEGG webpage
        pathway url will also be created \cr
     }
  }
  If you want to do GO functional analysis and KEGG pathway analysis for the
  following RNA-Seq workflow in R shell,
  please see \code{RNASeqGoKegg()} function.
}
\examples{
data(yeast)
\dontrun{
RNASeqGoKegg_CMD(RNASeqRParam  = yeast,
                 OrgDb.species = "org.Sc.sgd.db",
                 go.level = 3,
                 input.TYPE.ID = "GENENAME",
                 KEGG.organism = "sce")
}
}
\author{
Kuan-Hao Chao
}
