% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_pipline.R
\name{RNAseqPipeline}
\alias{RNAseqPipeline}
\title{rna seq pipline}
\usage{
RNAseqPipeline(path.prefix = "NOT_SET_YET",
  input.path.prefix = "NOT_SET_YET", gene.name = "NO_DATA",
  sample.pattern = "NO_DATA", num.parallel.threads = 8)
}
\description{
rna seq pipline
}
