% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_pipline.R
\name{RNAseqPipelineCMD}
\alias{RNAseqPipelineCMD}
\title{Create 'RNAseqPipline.R' that user can}
\usage{
RNAseqPipelineCMD(RNASeqWorkFlowParam, num.parallel.threads = 8)
}
\description{
Create 'RNAseqPipline.R' that user can
}
