% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_quality_assessment.R
\name{RNAseqQualityAssessment_CMD}
\alias{RNAseqQualityAssessment_CMD}
\title{Quality assessment of '.fastq.gz' files for RNA-Seq workflow in background.}
\usage{
RNAseqQualityAssessment_CMD(RNASeqWorkFlowParam, run = TRUE,
  check.s4.print = TRUE)
}
\arguments{
\item{RNASeqWorkFlowParam}{S4 object instance of experiment-related parameters}

\item{run}{Default value is \code{TRUE}. If \code{TRUE}, 'Rscript/Environment_Set.R' will be created and executed. The output log will be stored in 'Rscript_out/Environment_Set.Rout'.
If \code{False}, 'Rscript/Environment_Set.R' will be created without executed.}

\item{check.s4.print}{Default \code{TRUE}. If \code{TRUE}, the result of checking \code{RNASeqWorkFlowParam} will be reported in 'Rscript_out/Environment_Set.Rout'. If \code{FALSE}, the result of checking \code{RNASeqWorkFlowParam} will not be in 'Rscript_out/Environment_Set.Rout'}
}
\value{
None
}
\description{
Assess the quality of '.fastq.gz' files for RNA-Seq workflow in background. This step is optional in the whole RNA-Seq workflow.
This function reports the quality assessment result in two packages : \code{Rqc} and \code{systemPipeR}
For \code{Rqc}, 'RNAseq_results/QA_results/Rqc/Rqc_report.html' will be created
For \code{systemPipeR}, 'RNAseq_results/QA_results/Rqc/systemPipeR/fastqReport.pdf' will be created
If you want to assess the quality of '.fastq.gz' files for the following RNA-Seq workflow in R shell, please see \code{RNAseqQualityAssessment()} function.
}
\examples{
\dontrun{
input_file_dir <- system.file(package = "RNASeqWorkflow", "exdata")
exp <- RNASeqWorkFlowParam(path.prefix = "/tmp/", input.path.prefix = input_file_dir, genome.name = "hg19", sample.pattern = "SRR[0-9]",
                           experiment.type = "two.group", main.variable = "treatment", additional.variable = "cell")
RNAseqQualityAssessment_CMD(RNASeqWorkFlowParam = exp)}
}
