% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_quality_trimming.R
\name{RNASeqQualityTrimming_CMD}
\alias{RNASeqQualityTrimming_CMD}
\title{RNASeqQualityTrimming_CMD}
\usage{
RNASeqQualityTrimming_CMD(RNASeqRParam, cum.error = 1,
  trimming.position = NA, reads.length.limit = 36, run = TRUE,
  check.s4.print = TRUE)
}
\arguments{
\item{RNASeqRParam}{S4 object instance of
experiment-related parameters}

\item{cum.error}{Default \code{1}.
Cut of threshold of cumulative probability of error per base.}

\item{trimming.position}{Default \code{NA}. If value is \code{NA}, trimming
points of all paired-end files will be calculated by CDF of error rate, while
they will all be the value of \code{trimming.position}.}

\item{reads.length.limit}{Default \code{36}.
The shortest base pair length of short reads}

\item{run}{Default value is \code{TRUE}.
If \code{TRUE}, 'Rscript/Environment_Set.R' will be created and executed.
The output log will be stored in 'Rscript_out/Environment_Set.Rout'.
If \code{False}, 'Rscript/Environment_Set.R'
will be created without executed.}

\item{check.s4.print}{Default \code{TRUE}.
If \code{TRUE}, the result of checking \code{RNASeqRParam}
will be reported in 'Rscript_out/Environment_Set.Rout'.
If \code{FALSE}, the result of checking
\code{RNASeqRParam} will not be in
'Rscript_out/Environment_Set.Rout'}
}
\value{
None
}
\description{
Trim '.fastq.gz' files for RNA-Seq workflow in background.
  This step is optional in the whole RNA-Seq workflow.
  The trimming method is implemented by R package \code{ShortRead} \cr
  If you want to trim '.fastq.gz' files for the RNA-Seq workflow in R shell,
  please see \code{RNASeqQualityTrimming()} function.
}
\examples{
data(yeast)
\dontrun{
RNASeqQualityTrimming_CMD(RNASeqRParam = yeast)}
}
\author{
Kuan-Hao Chao
}
