% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_raw_read_process.R
\name{RNASeqReadProcess_CMD}
\alias{RNASeqReadProcess_CMD}
\title{RNASeqReadProcess_CMD}
\usage{
RNASeqReadProcess_CMD(RNASeqWorkFlowParam, num.parallel.threads = 1,
  run = TRUE, check.s4.print = TRUE, Hisat2.Index.run = TRUE,
  Hisat2.Alignment.run = TRUE, Samtools.Bam.run = TRUE,
  StringTie.Assemble.run = TRUE, StringTie.Merge.Trans.run = TRUE,
  Gffcompare.Ref.Sample.run = TRUE, StringTie.Ballgown.run = TRUE,
  PreDECountTable.run = TRUE)
}
\arguments{
\item{RNASeqWorkFlowParam}{S4 object instance of experiment-related
parameters}

\item{num.parallel.threads}{Specify the number of processing threads (CPUs)
to use for transcript assembly. The default is 1.}

\item{run}{Default value is \code{TRUE}. If \code{TRUE},
'Rscript/Environment_Set.R' will be created and executed.
The output log will be stored in 'Rscript_out/Environment_Set.Rout'.
If \code{False}, 'Rscript/Environment_Set.R' will be
created without executed.}

\item{check.s4.print}{Default \code{TRUE}. If \code{TRUE},
the result of checking \code{RNASeqWorkFlowParam}
will be reported in 'Rscript_out/Environment_Set.Rout'. If \code{FALSE},
the result of checking \code{RNASeqWorkFlowParam} will not be in
'Rscript_out/Environment_Set.Rout'.}

\item{Hisat2.Index.run}{Whether to run 'HISAT2 index' step in this function
step. Default value is \code{TRUE}. Set \code{FALSE} to skip
'HISAT2 index' step.}

\item{Hisat2.Alignment.run}{Whether to run 'HISAT2 alignment' step in this
function step. Default value is \code{TRUE}.
Set \code{FALSE} to skip 'HISAT2 alignment' step.}

\item{Samtools.Bam.run}{Whether to run 'SAMTools SAM to BAM' step in this
function step. Default value is \code{TRUE}.
Set \code{FALSE} to skip 'SAMTools SAM to BAM' step.}

\item{StringTie.Assemble.run}{Whether to run 'StringTie assembly' step in
this function step. Default value is \code{TRUE}.
Set \code{FALSE} to skip 'StringTie assembly' step.}

\item{StringTie.Merge.Trans.run}{Whether to run 'StringTie GTF merging' step
in this function step. Default value is \code{TRUE}.
Set \code{FALSE} to skip 'StringTie GTF merging' step.}

\item{Gffcompare.Ref.Sample.run}{Whether to run 'Gffcompare comparison' step
in this function step. Default value is \code{TRUE}.
Set \code{FALSE} to skip 'Gffcompare comparison' step.}

\item{StringTie.Ballgown.run}{Whether to run 'StringTie ballgown creation'
step in this function step. Default value is \code{TRUE}.
Set \code{FALSE} to skip 'StringTie ballgown creation' step.}

\item{PreDECountTable.run}{Whether to run 'gene raw reads count creation'
step in this function step. Default value is \code{TRUE}.
Set \code{FALSE} to skip 'gene raw reads count creation' step.}
}
\value{
None
}
\description{
Process raw reads for RNA-Seq workflow in background. \cr
  This function do 5 things : \cr
  \enumerate{
    \item 'Hisat2' : aligns raw reads to reference genome.
      If \code{indices.optional} in \code{RNASeqWorkFlowParam} is
      \code{FALSE}, Hisat2 indices will be created.\cr
    \item 'Samtools': converts '.sam' files to '.bam' files.\cr
    \item 'Stringtie': assembles alignments into transcript.\cr
    \item 'Gffcompare': examines how transcripts compare with the
      reference annotation. \cr
    \item 'Stringtie': creates input files for ballgown, edgeR and DESeq2.\cr
    \item raw reads count: create raw reads count for DESeq2 and edgeR \cr
  }
  Before running this function, \code{RNASeqEnvironmentSet_CMD()} or
  \code{RNASeqEnvironmentSet()} must be executed successfully. \cr
  If you want to process raw reads for the following RNA-Seq workflow
  in R shell, please see \code{RNASeqReadProcess()} function.\cr
}
\examples{
data(yeast)
\dontrun{
## Before run this function, make sure \\code{RNASeqEnvironmentSet_CMD()}
## (or\\code{RNASeqEnvironmentSet()}) is executed successfully.
RNASeqReadProcess_CMD(RNASeqWorkFlowParam = exp,
                      num.parallel.threads = 10)}
}
\author{
Kuan-Hao Chao
}
