% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{RNASeqRParam-constructor}
\alias{RNASeqRParam-constructor}
\alias{RNASeqRParam}
\title{RNASeqRParam}
\usage{
RNASeqRParam(path.prefix = NA, input.path.prefix = NA,
  genome.name = NA, sample.pattern = NA, independent.variable = NA,
  case.group = NA, control.group = NA, fastq.gz.type = NA)
}
\arguments{
\item{path.prefix}{Path prefix of 'gene_data/', 'RNASeq_bin/',
'RNASeq_results/', 'Rscript/' and 'Rscript_out/' directories.}

\item{input.path.prefix}{Path prefix of 'input_files/' directory.}

\item{genome.name}{variable of genome name defined in this RNA-Seq workflow
(ex. \code{genome.name}.fa, \code{genome.name}.gtf).}

\item{sample.pattern}{Regular expression of paired-end fastq.gz files under
'input_files/raw_fastq.gz'. Expression not includes \code{_[1,2].fastq.gz}.}

\item{independent.variable}{Independent variable for the biological
experiment design of two-group RNA-Seq workflow.}

\item{case.group}{Group name of the case group.}

\item{control.group}{Group name of the control group.}

\item{fastq.gz.type}{Specify the fastq.gz file type. 'PE' represents 
paired-end and 'SE' represents single-end.}
}
\value{
an object of class \code{RNASeqRParam}
}
\description{
Constructor function for RNASeqRParam objects
}
\examples{
input_files.path <- system.file("extdata/", package = "RNASeqRData")
rnaseq_result.path <- "/Users/chaokuan-hao/TRY"
# rnaseq_result.path <- tempdir(check = TRUE)
exp <- RNASeqRParam(path.prefix          = rnaseq_result.path,
                    input.path.prefix    = input_files.path,
                    genome.name          = "Saccharomyces_cerevisiae_XV_Ensembl",
                    sample.pattern       = "SRR[0-9]*_XV",
                    independent.variable = "state",
                    case.group           = "60mins_ID20_amphotericin_B",
                    control.group        = "60mins_ID20_control",
                    fastq.gz.type        = "PE")
}
\author{
kuan-hao Chao

Kuan-Hao Chao
}
