% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdditionalClasses_bam.R
\name{RNASeqRParam_Bam-constructor}
\alias{RNASeqRParam_Bam-constructor}
\alias{RNASeqRParam_Bam}
\title{RNASeqR_Bam}
\usage{
RNASeqRParam_Bam(path.prefix = NA, input.path.prefix = NA,
  genome.name = NA, sample.pattern = NA, independent.variable = NA,
  case.group = NA, control.group = NA)
}
\arguments{
\item{path.prefix}{Path prefix of 'gene_data/', 'RNASeq_bin/',
'RNASeq_results/', 'Rscript/' and 'Rscript_out/' directories.}

\item{input.path.prefix}{Path prefix of 'input_files/' directory.}

\item{genome.name}{variable of genome name defined in this RNA-Seq workflow
(ex. \code{genome.name}.fa, \code{genome.name}.gtf).}

\item{sample.pattern}{Regular expression of paired-end fastq.gz files under
'input_files/raw_bam'. Expression not includes \code{_[1,2].fastq.gz}.}

\item{independent.variable}{Independent variable for the biological
experiment design of two-group RNA-Seq workflow.}

\item{case.group}{Group name of the case group.}

\item{control.group}{Group name of the control group.}
}
\value{
an object of class \code{RNASeqRParam_Bam}
}
\description{
Constructor function for RNASeqRParam_Bam objects
}
\examples{
input_files.path <- system.file("extdata/", package = "RNASeqRData")
rnaseq_result.path <- tempdir(check = TRUE)
\dontrun{
exp <- RNASeqRParam_Bam(path.prefix          = rnaseq_result.path,
                        input.path.prefix    = input_files.path,
                        genome.name          = "Saccharomyces_cerevisiae_XV_Ensembl",
                        sample.pattern       = "SRR[0-9]*_XV",
                        independent.variable = "state",
                        case.group           = "60mins_ID20_amphotericin_B",
                        control.group        = "60mins_ID20_control")
}
}
\author{
Kuan-Hao Chao

Kuan-Hao Chao
}
