% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat2_samtools_stringtie_ballgown_related.R
\name{SamtoolsToBam}
\alias{SamtoolsToBam}
\title{use 'samtools' to sort and convert the SAM files to BAM}
\usage{
SamtoolsToBam(path.prefix, gene.name, sample.pattern,
  num.parallel.threads = 8)
}
\description{
use 'samtools' to sort and convert the SAM files to BAM
}
