% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/rnaseq_raw_read_process_hisat2_samtools_stringtie_ballgown_related.R
\name{StringTieAssemble}
\alias{StringTieAssemble}
\title{stringtie assemble and quantify expressed genes and transcripts}
\usage{
StringTieAssemble(path.prefix, genome.name, sample.pattern,
  num.parallel.threads = 8)
}
\description{
stringtie assemble and quantify expressed genes and transcripts
}
