% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/rnaseq_raw_read_process_hisat2_samtools_stringtie_ballgown_related.R
\name{StringTieReEstimate}
\alias{StringTieReEstimate}
\title{stringtie re-estimate the abundance}
\usage{
StringTieReEstimate(path.prefix, gene.name, sample.pattern,
  num.parallel.threads = 8)
}
\description{
stringtie re-estimate the abundance
}
