% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/rnaseq_raw_read_process_hisat2_samtools_stringtie_ballgown_related.R
\name{StringTieToBallgown}
\alias{StringTieToBallgown}
\title{stringtie estimate transcript abundances and create table counts for Ballgown}
\usage{
StringTieToBallgown(path.prefix, gene.name, sample.pattern,
  num.parallel.threads = 8)
}
\description{
stringtie estimate transcript abundances and create table counts for Ballgown
}
