% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_fastq_gz.R
\name{Update_Fastq_gz}
\alias{Update_Fastq_gz}
\title{Update_Fastq_gz}
\usage{
Update_Fastq_gz(RNASeqRParam, prepared_fastq_gz, target_samples = "ALL")
}
\arguments{
\item{RNASeqRParam}{S4 object instance of experiment-related
parameters}

\item{prepared_fastq_gz}{absolute path to the prepared 'raw_fastq.gz' 
directory.}

\item{target_samples}{list of samples that are going to update. Default
value is \code{ALL}}
}
\value{
None
}
\description{
This function let users update their trimmed fastq.gz files automatically.
}
\examples{
data(yeast)
\dontrun{
RNASeqDifferentialAnalysis(RNASeqRParam = yeast)}
}
\author{
Kuan-Hao Chao
}
