% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast-data.R
\docType{data}
\name{yeast}
\alias{yeast}
\title{Toy RNASeqRParam object}
\description{
Small RNASeqRParam S4 object created with checked valid
  parameters for demonstration purposes
}
\examples{
data(yeast)
yeast
# RNASeqRParam S4 object for example demonstration.
}
\author{
Kuan-Hao Chao
}
