% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat2_samtools_stringtie_ballgown_related.R
\name{BallgownPreprocess}
\alias{BallgownPreprocess}
\title{Run ballgown analysis}
\usage{
BallgownPreprocess(path.prefix, gene.name, sample.pattern)
}
\description{
Run ballgown analysis
}
