% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ballgown_analysis.R
\name{BallgownPreprocess}
\alias{BallgownPreprocess}
\title{Run ballgown analysis}
\usage{
BallgownPreprocess(path.prefix, gene.name, sample.pattern, independent.variable,
  ballgown.pval, ballgown.log2FC)
}
\description{
Run ballgown analysis
}
