% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ballgown_analysis.R
\name{BallgownVolcanoPlot}
\alias{BallgownVolcanoPlot}
\title{inner function : DEG volcanplot}
\usage{
BallgownVolcanoPlot(ballgown.pval = 0.05, ballgown.log2FC = 1)
}
\description{
inner function : DEG volcanplot
}
