% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ballgown_analysis.R
\name{BallgownVolcanoPlot}
\alias{BallgownVolcanoPlot}
\title{inner function : DEG volcanplot}
\usage{
BallgownVolcanoPlot(path.prefix, ballgown.pval, ballgown.log2FC)
}
\description{
inner function : DEG volcanplot
}
