% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_install_tools.R
\name{CheckToolAll}
\alias{CheckToolAll}
\title{Check necessary programs for this pipeline}
\usage{
CheckToolAll(print = TRUE)
}
\description{
Check whether Hisat2, Stringtie, Samtools, Gffcompare are installed on the workstation
}
