% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_environment_set.R
\name{CheckToolAll}
\alias{CheckToolAll}
\title{Check 'Hisat2', 'StringTie', 'Samtools' and 'Gffcompare' for this workflow}
\usage{
CheckToolAll(print = TRUE)
}
\arguments{
\item{print}{If \code{TRUE}, detailed information will be printed. If \code{FALSE}, detailed information will not be printed.}
}
\value{
None
}
\description{
Check whether 'Hisat2', 'Stringtie', 'Samtools' and 'Gffcompare' are installed on the workstation
}
