% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/rnaseq_raw_read_process_hisat2_samtools_stringtie_ballgown_related.R
\name{GffcompareRefSample}
\alias{GffcompareRefSample}
\title{Examine how the transcripts compare with the reference annotation}
\usage{
GffcompareRefSample(path.prefix, genome.name, sample.pattern)
}
\description{
Examine how the transcripts compare with the reference annotation
}
