% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/rnaseq_raw_read_process_hisat2_samtools_stringtie_ballgown_related.R
\name{Hisat2AlignmentDefault}
\alias{Hisat2AlignmentDefault}
\title{hisat2 alignment default}
\usage{
Hisat2AlignmentDefault(path.prefix, genome.name, sample.pattern,
  num.parallel.threads = 8)
}
\description{
hisat2 alignment default
}
