% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/rnaseq_raw_read_process_hisat2_samtools_stringtie_ballgown_related.R
\name{Hisat2ReportAssemble}
\alias{Hisat2ReportAssemble}
\title{Report Hisat2 assemble rate}
\usage{
Hisat2ReportAssemble(path.prefix, genome.name, sample.pattern)
}
\description{
Report Hisat2 assemble rate
}
