% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepDE_create_count_table.R
\name{PreDECountTable}
\alias{PreDECountTable}
\title{converting stringtie ballogwn preprocessed data to count table}
\usage{
PreDECountTable(path.prefix, sample.pattern, python.variable, print = TRUE)
}
\description{
converting stringtie ballogwn preprocessed data to count table
}
