% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_differential_analysis.R
\name{RNASeqDifferentialAnalysis}
\alias{RNASeqDifferentialAnalysis}
\title{RNASeqDifferentialAnalysis}
\usage{
RNASeqDifferentialAnalysis(RNASeqRParam, which.trigger = "OUTSIDE",
  INSIDE.path.prefix = NA, ballgown.run = TRUE, ballgown.pval = 0.05,
  ballgown.log2FC = 1, DESeq2.run = TRUE, DESeq2.pval = 0.1,
  DESeq2.log2FC = 1, edgeR.run = TRUE, edgeR.pval = 0.05,
  edgeR.log2FC = 1, check.s4.print = TRUE)
}
\arguments{
\item{RNASeqRParam}{S4 object instance of experiment-related
parameters}

\item{which.trigger}{Default value is \code{OUTSIDE}. User should not change
this value.}

\item{INSIDE.path.prefix}{Default value is \code{NA}. User should not change
this value.}

\item{ballgown.run}{Default \code{TRUE}. Logical value whether to run
ballgown differential analysis.}

\item{ballgown.pval}{Default \code{0.05}. Set the threshold of ballgown
p-value to filter out differential expressed gene.}

\item{ballgown.log2FC}{Default \code{1}. Set the threshold of ballgown
log2 fold change to filter out differential expressed gene.}

\item{DESeq2.run}{Default \code{TRUE}. Logical value whether to run
DESeq2 differential analysis.}

\item{DESeq2.pval}{Default \code{0.05}. Set the threshold of DESeq2 p-value
to filter out differential expressed gene.}

\item{DESeq2.log2FC}{Default \code{1}. Set the threshold of DESeq2 log2
fold change to filter out differential expressed gene.}

\item{edgeR.run}{Default \code{TRUE}. Logical value whether to run
edgeR differential analysis.}

\item{edgeR.pval}{Default \code{0.05}. Set the threshold of edgeR p-value
to filter out differential expressed gene.}

\item{edgeR.log2FC}{Default \code{1}. Set the threshold of edgeR log2
fold change to filter out differential expressed gene.}

\item{check.s4.print}{Default \code{TRUE}. If \code{TRUE}, the result of
checking \code{RNASeqRParam} will be reported in
'Rscript_out/Environment_Set.Rout'. If \code{FALSE}, the result of checking
\code{RNASeqRParam} will not be in
'Rscript_out/Environment_Set.Rout'.}
}
\value{
None
}
\description{
This function will run differential analysis on ballgown,
  DESeq2 and edgeR in background. \cr
  This function do following things : \cr
  \enumerate{
    \item ballgown analysis \cr
    Raw reads are normalized into FPKM values \cr
    The main statistic test in ballgown is paramatic F-test comparing nested
    linear models \cr
    \item DESeq2 analysis  \cr
     Median of rations normalization(MRN) is used in DESeq2 for raw reads
      count normalization.  \cr
     Sequencing depth and RNA composition is taken into consideration is this
      normalization method.  \cr
     The main statistic test in DESeq2 is negative binomial distribution. \cr
    \item edgeR analysis  \cr
     Raw reads are normalized by TMM and library size.
      (run \code{calcNormFactors()} to get a DGEList,
     and then run \code{cpm()} on that DGEList)  \cr
     The main statistic test in edgeR is trimmed mean of M-values(TMM).\cr
  }
  If you want to run differential analysis on ballgown,
  DESeq2, edgeR for the following RNA-Seq workflow in background,
  please see \code{RNASeqDifferentialAnalysis()} function.
}
\examples{
data(yeast)
\dontrun{
RNASeqDifferentialAnalysis(RNASeqRParam = yeast)}
}
\author{
Kuan-Hao Chao
}
