% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_differential_analysis.R
\name{RNASeqDifferentialAnalysis}
\alias{RNASeqDifferentialAnalysis}
\title{Ballgown analysis, TPM & Student's t-test analysis, DESeq2 analysis, edgeR analysis for RNA-Seq workflow in R shell}
\usage{
RNASeqDifferentialAnalysis(path.prefix, genome.name, sample.pattern,
  independent.variable, control.group, experiment.group,
  ballgown.pval = 0.05, ballgown.log2FC = 1, TPM.pval = 0.05,
  TPM.log2FC = 1, DESeq2.pval = 0.1, DESeq2.log2FC = 1,
  edgeR.pval = 0.05, edgeR.log2FC = 1)
}
\arguments{
\item{path.prefix}{path prefix of 'gene_data/', 'RNASeq_bin/', 'RNASeq_results/', 'Rscript/' and 'Rscript_out/' directories}

\item{genome.name}{genome.name Variable of genome name defined in this RNA-Seq workflow (ex. \code{genome.name}.fa, \code{genome.name}.gtf)}

\item{sample.pattern}{sample.pattern  Regular expression of paired-end fastq.gz files under 'input_files/raw_fastq.gz'. Expression not includes \code{_[1,2].fastq.gz}.}

\item{independent.variable}{independent variable for the biological experiment design of two-group RNA-Seq workflow}

\item{control.group}{group name of the control group}

\item{experiment.group}{group name of the experiment group}

\item{ballgown.pval}{Default \code{0.05}. Set the threshold of ballgown p-value to filter out differential expressed gene.}

\item{ballgown.log2FC}{Default \code{1}. Set the threshold of ballgown log2 fold change to filter out differential expressed gene.}

\item{TPM.pval}{Default \code{0.05}. Set the threshold of TPM & student t-test p-value to filter out differential expressed gene.}

\item{TPM.log2FC}{Default \code{1}. Set the threshold of TPM & student t-test log2 fold change to filter out differential expressed gene.}

\item{DESeq2.pval}{Default \code{0.05}. Set the threshold of DESeq2 p-value to filter out differential expressed gene.}

\item{DESeq2.log2FC}{Default \code{1}. Set the threshold of DESeq2 log2 fold change to filter out differential expressed gene.}

\item{edgeR.pval}{Default \code{0.05}. Set the threshold of edgeR p-value to filter out differential expressed gene.}

\item{edgeR.log2FC}{Default \code{1}. Set the threshold of edgeR log2 fold change to filter out differential expressed gene.}
}
\value{
None
}
\description{
This function will run differential analysis on ballgown, DESeq2 and edgeR. TPM values will be converted from ballgown FPKM values, and T-test will be made to calculate pvalue.
It is strongly advised to run \code{RNASeqDifferentialAnalysis_CMD()} directly. Running \code{RNASeqDifferentialAnalysis_CMD()} will create 'Differential_Analysis.Rout' file in 'Rscript_out/' directory.
This function do following things :
1. ballgown analysis
   Raw reads are normalized into FPKM values
   The main statistic test in ballgown is paramatic F-test comparing nested linear models
2. TPM & Student's t-test analysis
   TPM normalization is calculated from the FPKM values produced by ballgown.
   Independent sample t-test is used with TPM values as input. P-values will be produced after t-test. Fold change values are calculated by the formula : "average TPM experiment group values divide average TPM control group values".
3. DESeq2 analysis
   Median of rations normalization(MRN) is used in DESeq2 for raw counts normalization. Sequencing depth and RNA composition is taken into consideration is this normalization method.
   The main statistic test in DESeq2 is negative binomial distribution.
4. edgeR analysis
   Raw reads are normalized by TMM and library size. (run \code{calcNormFactors()} to get a DGEList, and then run \code{cpm()} on that DGEList)
   The main statistic test in edgeR is trimmed mean of M-values(TMM).
If you want to run differential analysis on ballgown, TPM normalization, DESeq2, edgeR for the following RNA-Seq workflow in background, please see \code{RNASeqDifferentialAnalysis()} function.
}
\examples{
\dontrun{
input_file_dir <- system.file(package = "RNASeqWorkflow", "exdata")
exp <- RNASeqWorkFlowParam(path.prefix = "/tmp/", input.path.prefix = input_file_dir, genome.name = "hg19", sample.pattern = "SRR[0-9]",
                           experiment.type = "two.group", main.variable = "treatment", additional.variable = "cell")
RNASeqEnvironmentSet_CMD(RNASeqWorkFlowParam = exp)}
}
\author{
Kuan-Hao Chao
}
