% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_environment_set.R
\name{RNAseqEnvironmentSet}
\alias{RNAseqEnvironmentSet}
\title{Set up the environment for the RNA seq pipeline in R shell}
\usage{
RNAseqEnvironmentSet(path.prefix, input.path.prefix, genome.name,
  sample.pattern, indexes.optional, os.type)
}
\arguments{
\item{path.prefix}{the directory holding installations and analysis results}

\item{input.path.prefix}{user has to prepared valid 'input_files/' under this directory}

\item{genome.name}{gene name defined in this RNA-Seq workflow (ex. genome.name.fa, genome.name.gtf)}

\item{sample.pattern}{sample pattern describing the name of raw fastq.gz files}

\item{indexes.optional}{logical value whether indexes/ is exit in 'input_files/'}

\item{os.type}{'linux' or 'osx'. The operating system type}
}
\description{
1. To set up the environment for the following RNA seq pipeline in R shell, create 'RNASeqWorkFlowParam' first, get the value from the instance of this S4 object, and then put set the corresponding value to the input of this function.
2. To set up the environment for the following RNA seq pipeline in background, create 'RNASeqWorkFlowParam' first, and run 'RNAseqEnvironmentSet_CMD(RNASeqWorkFlowParam)'.
It is recommended to set up the environment in background.
}
