% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_install_tools.R
\name{RNAseqEnvironmentSet}
\alias{RNAseqEnvironmentSet}
\title{Set up the environment for the RNA seq pipeline in R shell}
\usage{
RNAseqEnvironmentSet(path.prefix, input.path.prefix, gene.name, sample.pattern,
  optional, os.type)
}
\description{
1. To set up the environment for the following RNA seq pipeline in R shell, create 'RNASeqWorkFlowParam' first, get the value from the instance of this S4 object, and then put set the corresponding value to the input of this function.
2. To set up the environment for the following RNA seq pipeline in background, create 'RNASeqWorkFlowParam' first, and run 'RNAseqEnvironmentSet_CMD(RNASeqWorkFlowParam)'.
It is recommended to set up the environment in background.
}
