% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_environment_set.R
\name{RNAseqEnvironmentSet}
\alias{RNAseqEnvironmentSet}
\title{Environment setting for RNA-Seq workflow in R shell}
\usage{
RNAseqEnvironmentSet(path.prefix, input.path.prefix, genome.name,
  sample.pattern, indexes.optional, os.type, mkdir.bool = TRUE)
}
\arguments{
\item{path.prefix}{path prefix of 'gene_data/', 'RNAseq_bin/', 'RNAseq_results/', 'Rscript/' and 'Rscript_out/' directories}

\item{input.path.prefix}{path prefix of 'input_files/' directory}

\item{genome.name}{variable of genome name defined in this RNA-Seq workflow (ex. genome.name.fa, genome.name.gtf)}

\item{sample.pattern}{regular expression of raw fastq.gz files under 'input_files/raw_fastq.gz'}

\item{indexes.optional}{logical value whether indexes/ is exit in 'input_files/'}

\item{os.type}{'linux' or 'osx'. The operating system type}

\item{mkdir.bool}{Default \code{TRUE}. If \code{TRUE}, environment directories will be created. If \code{FALSE}, no directories will be created. When executing RNAseqEnvironmentSet(), 'mkdir.bool' should always be \code{TRUE}}
}
\value{
None
}
\description{
Set up the environment for the following RNA-Seq workflow in R shell.
It is strongly advised to run \code{RNAseqEnvironmentSet_CMD()} directly. Running this function directly is not recommended.
This function do 4 things : 1. Create file directories. 2. Install necessary tools. 3. Export 'RNAseq_bin/' to the R environment. 4. Check command of tools.
First it will create 'gene_data/', 'RNAseq_bin/', 'RNAseq_results/', 'Rscript/', 'Rscript_out/' directories. Afterwards, 'Hisat2', 'Stringtie', 'Samtools',
'Gffcompare' will be installed under 'RNAseq_bin/'. 'RNAseq_bin/' will be added to the R environment and validity of tools will be checked. Any ERROR occurs will be reported and the program will be terminated.
If you want to set up the environment for the following RNA-Seq workflow in background, please see \code{RNAseqEnvironmentSet_CMD()} function.
}
\examples{
\dontrun{
exp <- RNASeqWorkFlowParam(path.prefix = "/home/rnaseq", input.path.prefix = "/home", genome.name = "hg19", sample.pattern = "SRR[0-9]",
                           independent.variable = "two.group", control.group = "treatment", experiment.group = "cell")
RNAseqEnvironmentSet(path.prefix = exp@path.prefix, input.path.prefix = exp@input.path.prefix, genome.name = exp@genome.name,
                     sample.pattern = exp@sample.pattern, indexes.optional = exp@indexes.optional, os.type = exp@os.type)}
}
