% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_environment_set.R
\name{RNAseqEnvironmentSet_CMD}
\alias{RNAseqEnvironmentSet_CMD}
\title{Set up the environment for the following RNA seq pipeline in background.}
\usage{
RNAseqEnvironmentSet_CMD(RNASeqWorkFlowParam, run = TRUE,
  check.s4.print = TRUE)
}
\arguments{
\item{RNASeqWorkFlowParam}{S4 object instance of experiment-related parameters}
}
\description{
To set up the environment for the following RNA seq pipeline in background, create 'RNASeqWorkFlowParam' first, and then run 'RNAseqEnvironmentSet_CMD(RNASeqWorkFlowParam)'.
If you want to set up the environment for the following RNA seq pipeline in R shell, please see 'RNAseqEnvironmentSet()' function.
This function do 4 things : 1. Create file directories. 2. Install necessary tools. 3. Export 'RNAseq_bin/' to the R environment. 4. Check command of tools.
First it will create 'gene_data/', 'RNAseq_bin/', 'RNAseq_results/', 'Rscript/', 'Rscript_out/' directories. Afterwards, 'Hisat2', 'Stringtie', 'Samtools',
'Gffcompare' will be installed under 'RNAseq_bin/'. 'RNAseq_bin/' will be added to the R environment and validity of tools will be checked. Any ERROR occurs will be reported.
}
