% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_environment_set.R
\name{RNAseqEnvironmentSet_CMD}
\alias{RNAseqEnvironmentSet_CMD}
\title{Environment setting for RNA-Seq workflow in background}
\usage{
RNAseqEnvironmentSet_CMD(RNASeqWorkFlowParam, run = TRUE,
  check.s4.print = TRUE)
}
\arguments{
\item{RNASeqWorkFlowParam}{S4 object instance of experiment-related parameters}

\item{run}{Default value is \code{TRUE}. If \code{TRUE}, 'Rscript/Environment_Set.R' will be created and executed. The output log will be stored in 'Rscript_out/Environment_Set.Rout'.
If \code{False}, 'Rscript/Environment_Set.R' will be created without executed.}

\item{check.s4.print}{Default \code{TRUE}. If \code{TRUE}, the result of checking \code{RNASeqWorkFlowParam} will be reported in 'Rscript_out/Environment_Set.Rout'. If \code{FALSE}, the result of checking \code{RNASeqWorkFlowParam} will not be in 'Rscript_out/Environment_Set.Rout'}

\item{mkdir.bool}{Default \code{TRUE}. If \code{TRUE}, environment directories will be created. If \code{FALSE}, no directories will be created. When executing RNAseqEnvironmentSet(), 'mkdir.bool' should always be \code{TRUE}}
}
\value{
None
}
\description{
Set up the environment for the following RNA-Seq workflow in background.
This function do 4 things :
1. Create file directories.
2. Install necessary tools.
3. Export 'RNAseq_bin/' to the R environment.
4. Check command of tools.
First it will create 'gene_data/', 'RNAseq_bin/', 'RNAseq_results/', 'Rscript/', 'Rscript_out/' directories. Afterwards, 'Hisat2', 'Stringtie', 'Samtools',
'Gffcompare' will be installed under 'RNAseq_bin/'. 'RNAseq_bin/' will be added to the R environment and validity of tools will be checked. Any ERROR occurs will be reported and the program will be terminated.
If you want to set up the environment for the following RNA-Seq workflow in R shell, please see \code{RNAseqEnvironmentSet()} function.
}
\examples{
\dontrun{
input_file_dir <- system.file(package = "RNASeqWorkflow", "exdata")
exp <- RNASeqWorkFlowParam(path.prefix = "/tmp/", input.path.prefix = input_file_dir, genome.name = "hg19", sample.pattern = "SRR[0-9]",
                           experiment.type = "two.group", main.variable = "treatment", additional.variable = "cell")
RNAseqEnvironmentSet_CMD(RNASeqWorkFlowParam = exp)}
}
