% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_GO_KEGG.R
\name{RNASeqGoKegg}
\alias{RNASeqGoKegg}
\title{RNASeqGoKegg}
\usage{
RNASeqGoKegg(RNASeqRParam, which.trigger = "OUTSIDE",
  INSIDE.path.prefix = NA, OrgDb.species, go.level = 3, input.TYPE.ID,
  KEGG.organism, check.s4.print = TRUE)
}
\arguments{
\item{RNASeqRParam}{S4 object instance of experiment-related
parameters}

\item{which.trigger}{Default value is \code{OUTSIDE}. User should not change
this value.}

\item{INSIDE.path.prefix}{Default value is \code{NA}. User should not change
this value.}

\item{OrgDb.species}{the genome wide annotation packages of species
on Bioconductor. Currently, there are 19 supported genome wide
annotation packages of species.}

\item{go.level}{the depth of acyclic graph in GO analysis}

\item{input.TYPE.ID}{The gene name type in OrgDb.species annotation packahge.}

\item{KEGG.organism}{the species that are supported for KEGG analysis.
Currently, there are more than 5000 supported species genome.
Check the valid species terms on
https://www.genome.jp/kegg/catalog/org_list.html}

\item{check.s4.print}{Default \code{TRUE}. If \code{TRUE},
the result of checking \code{RNASeqRParam} will be reported in
'Rscript_out/Environment_Set.Rout'. If \code{FALSE}, the result of checking
\code{RNASeqRParam} will not be in
'Rscript_out/Environment_Set.Rout'}
}
\value{
None
}
\description{
Run Gene Ontology(GO) and Kyoto Encyclopedia of Genes and Genomes(KEGG)
  analysis in R shell. \cr
  This function do Gene Ontology(GO) and Kyoto Encyclopedia of Genes and
  Genomes(KEGG) analysis : \cr
  \enumerate{
    \item Gene Ontology(GO) :\cr
     \enumerate{
       \item Do GO function classification analysis. \cr
       \item Do GO function enrichment analysis. \cr
       \item Visualization : bar plot, dot plot etc. \cr
     }
    \item Kyoto Encyclopedia of Genes and Genomes(KEGG) :\cr
     \enumerate{
       \item Do KEGG pathway enrichment analysis \cr
       \item Pathway visulization with \code{pathview} package. KEGG webpage
        pathway url will also be created \cr
     }
  }
  If you want to do GO functional analysis and KEGG pathway analysis
  for the following RNA-Seq workflow in background,
   please see \code{RNASeqGoKegg_CMD()} function.
}
\examples{
data(yeast)
\dontrun{
RNASeqGoKegg(RNASeqRParam  = yeast,
             OrgDb.species = "org.Sc.sgd.db",
             go.level = 3,
             input.TYPE.ID = "GENENAME",
             KEGG.organism = "sce")}
}
\author{
Kuan-Hao Chao
}
