% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_GO_KEGG.R
\name{RNASeqGoKegg}
\alias{RNASeqGoKegg}
\title{RNASeqGoKegg}
\usage{
RNASeqGoKegg(path.prefix, independent.variable, OrgDb.species,
  go.level = 3, input.TYPE.ID, KEGG.organism)
}
\arguments{
\item{path.prefix}{Path prefix of 'gene_data/', 'RNASeq_bin/',
'RNASeq_results/', 'Rscript/' and 'Rscript_out/' directories.}

\item{independent.variable}{independent variable for the biological
experiment design of two-group RNA-Seq workflow}

\item{OrgDb.species}{the genome wide annotation packages of species
on Bioconductor. Currently, there are 19 supported genome wide
annotation packages of species.}

\item{KEGG.organism}{the species that are supported for KEGG analysis.
Currently, there are more than 5000 supported species genome.
Check the valid species terms on
https://www.genome.jp/kegg/catalog/org_list.html}
}
\value{
None
}
\description{
Run Gene Ontology(GO) and Kyoto Encyclopedia of Genes and Genomes(KEGG)
  analysis in R shell. \cr
  This function do Gene Ontology(GO) and Kyoto Encyclopedia of Genes and
  Genomes(KEGG) analysis : \cr
  \enumerate{
    \item Gene Ontology(GO) :\cr
     \enumerate{
       \item Do GO function classification analysis. \cr
       \item Do GO function enrichment analysis. \cr
       \item Visualization : bar plot, dot plot etc. \cr
     }
    \item Kyoto Encyclopedia of Genes and Genomes(KEGG) :\cr
     \enumerate{
       \item Do KEGG pathway enrichment analysis \cr
       \item Pathway visulization with \code{pathview} package. KEGG webpage
        pathway url will also be created \cr
     }
  }
  If you want to do GO functional analysis and KEGG pathway analysis
  for the following RNA-Seq workflow in background,
   please see \code{RNASeqGoKegg_CMD()} function.
}
\examples{
data(yeast)
\dontrun{
RNASeqGoKegg(path.prefix          = path.prefix@yeast,
             independent.variable = independent.variable@yeast,
             OrgDb.species        = "org.Sc.sgd.db",
             go.level             = 3,
             input.TYPE.ID        = "GENENAME",
             KEGG.organism        = "sce")
}
}
\author{
Kuan-Hao Chao
}
