% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_GO_KEGG.R
\name{RNASeqGoKegg_CMD}
\alias{RNASeqGoKegg_CMD}
\title{GO functional analysis and KEGG pathway analysis for RNA-Seq workflow in background}
\usage{
RNASeqGoKegg_CMD(RNASeqWorkFlowParam, OrgDb.species, go.level = 3,
  input.TYPE.ID, KEGG.organism, run = TRUE, check.s4.print = TRUE)
}
\arguments{
\item{RNASeqWorkFlowParam}{S4 object instance of experiment-related parameters}

\item{OrgDb.species}{the genome wide annotation packages of species on Bioconductor. Currently, there are 19 supported genome wide annotation packages of species.}

\item{KEGG.organism}{the species that are supported for KEGG analysis. Currently, there are more than 5000 supported species genome. Check the valid species terms on https://www.genome.jp/kegg/catalog/org_list.html}

\item{run}{Default value is \code{TRUE}. If \code{TRUE}, 'Rscript/Environment_Set.R' will be created and executed. The output log will be stored in 'Rscript_out/Environment_Set.Rout'.
If \code{False}, 'Rscript/Environment_Set.R' will be created without executed.}

\item{check.s4.print}{Default \code{TRUE}. If \code{TRUE}, the result of checking \code{RNASeqWorkFlowParam} will be reported in 'Rscript_out/Environment_Set.Rout'. If \code{FALSE}, the result of checking \code{RNASeqWorkFlowParam} will not be in 'Rscript_out/Environment_Set.Rout'}
}
\value{
None
}
\description{
Run Gene Ontology(GO) and Kyoto Encyclopedia of Genes and Genomes(KEGG) analysis in background.
This function do Gene Ontology(GO) and Kyoto Encyclopedia of Genes and Genomes(KEGG) analysis :
1. Gene Ontology(GO) :
   All gene set :
       1. Do GO gene set enrichment analysis.
   Differential expressed gene set :
       1. Do GO function classification analysis.
       2. Do GO function enrichment analysis.
       3. Visualization : bar plot, dot plot etc.
2. Kyoto Encyclopedia of Genes and Genomes(KEGG) :
   All gene set :
       1. Do KEGG gene set enrichment analysis.
   Differential expressed gene set :
       1. Do KEGG pathway gene set enrichment analysis
       2. Do KEGG pathway enrichment analysis
       3. Pathway visulization with \code{pathview} package. KEGG webpage pathway url will also be created
If you want to do GO functional analysis and KEGG pathway analysis for the following RNA-Seq workflow in R shell, please see \code{RNASeqGoKegg()} function.
}
\examples{
data(yeast)
\dontrun{
RNASeqGoKegg_CMD(yeast,
                 OrgDb.species = "org.Sc.sgd.db",
                 go.level = 3,
                 input.TYPE.ID = "GENENAME",
                 KEGG.organism = "sce")
}
}
\author{
Kuan-Hao Chao
}
