% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_pipline.R
\name{RNAseqPipeline}
\alias{RNAseqPipeline}
\title{rna seq pipline}
\usage{
RNAseqPipeline(path.prefix, input.path.prefix, gene.name, sample.pattern,
  num.parallel.threads = 8, indexes.optional)
}
\description{
rna seq pipline
}
