% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_pipline.R
\name{RNAseqPipeline_CMD}
\alias{RNAseqPipeline_CMD}
\title{Run RNA seq pipeline with default parameter}
\usage{
RNAseqPipeline_CMD(RNASeqWorkFlowParam, num.parallel.threads = 8)
}
\description{
Run RNA seq pipeline with default parameter
}
