% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_quality_assessment.R
\name{RNAseqQualityAssessment}
\alias{RNAseqQualityAssessment}
\title{Quality control}
\usage{
RNAseqQualityAssessment(path.prefix, input.path.prefix, sample.pattern)
}
\description{
Quality control
}
