% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_quality_trimming.R
\name{RNASeqQualityTrimming}
\alias{RNASeqQualityTrimming}
\title{RNASeqQualityTrimming}
\usage{
RNASeqQualityTrimming(RNASeqRParam, which.trigger = "OUTSIDE",
  INSIDE.path.prefix = NA, cum.error = 1, trimming.position = NA,
  reads.length.limit = 36, check.s4.print = TRUE)
}
\arguments{
\item{RNASeqRParam}{S4 object instance of experiment-related
parameters}

\item{which.trigger}{Default value is \code{OUTSIDE}. User should not change
this value.}

\item{INSIDE.path.prefix}{Default value is \code{NA}. User should not change
this value.}

\item{cum.error}{Default \code{1}.
Cut of threshold of cumulative probability of error per base.}

\item{trimming.position}{Default \code{NA}. If value is \code{NA}, trimming
points of all paired-end files will be calculated by CDF of error rate,
while they will all be the value of \code{trimming.position}.}

\item{reads.length.limit}{Default \code{36}.
The shortest base pair length of short reads}

\item{check.s4.print}{Default \code{TRUE}.
If \code{TRUE}, the result of checking \code{RNASeqRParam}
will be reported in 'Rscript_out/Environment_Set.Rout'.
If \code{FALSE}, the result of checking
\code{RNASeqRParam} will not be in
'Rscript_out/Environment_Set.Rout'}
}
\value{
None
}
\description{
Trim '.fastq.gz' files for RNA-Seq workflow in R shell.
  This step is optional in the whole RNA-Seq workflow.
  It is strongly advised to run \code{RNASeqQualityTrimming_CMD()} directly.
  Running \code{RNASeqQualityTrimming_CMD()} will create
  'Quality_Trimming.Rout' file in 'Rscript_out/' directory.
  The trimming method is implemented by R package \code{ShortReads} \cr
  If you want to trim '.fastq.gz' files for the RNA-Seq workflow
  in background, please see \code{RNASeqQualityTrimming_CMD()} function.
}
\examples{
data(yeast)
\dontrun{
RNASeqQualityTrimming(RNASeqRParam = yeast)}
}
\author{
Kuan-Hao Chao
}
