% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_quality_trimming.R
\name{RNAseqQualityTrimming}
\alias{RNAseqQualityTrimming}
\title{Sample trimming of '.fastq.gz' files for RNA-Seq workflow in R shell}
\usage{
RNAseqQualityTrimming(path.prefix, sample.pattern, truncateStartBases = 0,
  truncateEndBases = 0, complexity = NULL, minLength = 50, nBases = 2)
}
\arguments{
\item{path.prefix}{path prefix of 'gene_data/', 'RNAseq_bin/', 'RNAseq_results/', 'Rscript/' and 'Rscript_out/' directories}

\item{sample.pattern}{regular expression of raw fastq.gz files under 'input_files/raw_fastq.gz'}

\item{nBases}{}
}
\description{
Trim '.fastq.gz' files for RNA-Seq workflow in R shell. This step is optional in the whole RNA-Seq workflow.
The trimming method is implemented by R package \code{QuasR}
If you want to trim '.fastq.gz' files for the RNA-Seq workflow in background, please see \code{RNAseqQualityTrimming_CMD()} function.
}
\examples{
input_file_dir <- system.file(package = "RNASeqWorkflow", "exdata")
exp <- RNASeqWorkFlowParam(path.prefix = "/tmp/", input.path.prefix = input_file_dir, genome.name = "hg19", sample.pattern = "SRR[0-9]",
                           experiment.type = "two.group", main.variable = "treatment", additional.variable = "cell")
RNAseqEnvironmentSet(path.prefix = exp@path.prefix, sample.pattern = exp@sample.pattern)
}
