% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_quality_trimming.R
\name{RNAseqQualityTrimming}
\alias{RNAseqQualityTrimming}
\title{Sample trimming of '.fastq.gz' files for RNA-Seq workflow in R shell}
\usage{
RNAseqQualityTrimming(path.prefix, sample.pattern,
  truncateStartBases = 0, truncateEndBases = 0, complexity = NULL,
  minLength = 50, nBases = 2)
}
\arguments{
\item{path.prefix}{Path prefix of 'gene_data/', 'RNAseq_bin/', 'RNAseq_results/', 'Rscript/' and 'Rscript_out/' directories}

\item{sample.pattern}{Regular expression of raw fastq.gz files under 'input_files/raw_fastq.gz'}

\item{truncateStartBases}{Function parameter of \code{QuasR::preprocessReads()}. The number of bases to be truncated (removed) from the begining of each sequence.}

\item{truncateEndBases}{Function parameter of \code{QuasR::preprocessReads()}. The number of bases to be truncated (removed) from the end of each sequence.}

\item{complexity}{Function parameter of \code{QuasR::preprocessReads()}. \code{NULL} or numeric(1). Default value is \code{NULL}.
If not NULL, the minimal sequence complexity, as a fraction of the average complexity in the human genome (~3.9bits).
For example, complexity = 0.5 will filter out sequences that do not have at least half the complexity of the human genome.
See Details on how the complexity is calculated.}

\item{minLength}{Function parameter of \code{QuasR::preprocessReads()}. The minimal allowed sequence length.}

\item{nBases}{Function parameter of \code{QuasR::preprocessReads()}. The maximal number of Ns allowed per sequence.}
}
\value{
None
}
\description{
Trim '.fastq.gz' files for RNA-Seq workflow in R shell. This step is optional in the whole RNA-Seq workflow.
It is strongly advised to run \code{RNAseqQualityTrimming_CMD()} directly. Running this function directly is not recommended.
The trimming method is implemented by R package \code{QuasR}
If you want to trim '.fastq.gz' files for the RNA-Seq workflow in background, please see \code{RNAseqQualityTrimming_CMD()} function.
}
\examples{
\dontrun{
input_file_dir <- system.file(package = "RNASeqWorkflow", "exdata")
exp <- RNASeqWorkFlowParam(path.prefix = "/tmp/", input.path.prefix = input_file_dir, genome.name = "hg19", sample.pattern = "SRR[0-9]",
                           experiment.type = "two.group", main.variable = "treatment", additional.variable = "cell")
RNAseqEnvironmentSet(path.prefix = exp@path.prefix, sample.pattern = exp@sample.pattern)}
}
