% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_raw_read_process.R
\name{RNAseqRawReadProcess}
\alias{RNAseqRawReadProcess}
\title{rna seq pipline}
\usage{
RNAseqRawReadProcess(path.prefix, input.path.prefix, genome.name,
  sample.pattern, python.variable.answer, python.variable.version,
  num.parallel.threads = 8, indexes.optional)
}
\description{
rna seq pipline
}
