% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_pipline.R
\name{RNAseqRawReadProcess_CMD}
\alias{RNAseqRawReadProcess_CMD}
\title{Run RNA seq raw read processing (alignment, assembly, expressive expression) in background.}
\usage{
RNAseqRawReadProcess_CMD(RNASeqWorkFlowParam, num.parallel.threads = 8)
}
\arguments{
\item{RNASeqWorkFlowParam}{S4 object instance of experiment-related parameters}
}
\description{
To process RNA seq raw read in background, you have to run 'RNASeqWorkFlowParam()' to create S4 object beforehands.
If you want to process RNA seq raw read in R shell, please see 'RNAseqRawReadProcess()' function.
This function do 4 things : 1. Hisat2 : raw reads align to reference genome. 2. Stringtie : Alignments assembly into transcript. 3. Stringtie : Expressive gene and transcript quantification
4. Compare aligned genome with reference genome
}
