% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd_batch_rnaseq_raw_read_process.R
\name{RNASeqReadProcess_CMD}
\alias{RNASeqReadProcess_CMD}
\title{Raw reads process (alignment, assembly, expressive quantification) of RNA-Seq in background.}
\usage{
RNASeqReadProcess_CMD(RNASeqWorkFlowParam, num.parallel.threads = 1,
  run = TRUE, check.s4.print = TRUE, Hisat2.Index.run = TRUE,
  Hisat2.Alignment.run = TRUE, Samtools.Bam.run = TRUE,
  StringTie.Assemble.run = TRUE, StringTie.Merge.Trans.run = TRUE,
  Gffcompare.Ref.Sample.run = TRUE, StringTie.Ballgown.run = TRUE,
  PreDECountTable.run = TRUE)
}
\arguments{
\item{RNASeqWorkFlowParam}{S4 object instance of experiment-related parameters}

\item{num.parallel.threads}{Specify the number of processing threads (CPUs) to use for transcript assembly. The default is 1.}

\item{run}{Default value is \code{TRUE}. If \code{TRUE}, 'Rscript/Environment_Set.R' will be created and executed. The output log will be stored in 'Rscript_out/Environment_Set.Rout'.
If \code{False}, 'Rscript/Environment_Set.R' will be created without executed.}

\item{check.s4.print}{Default \code{TRUE}. If \code{TRUE}, the result of checking \code{RNASeqWorkFlowParam} will be reported in 'Rscript_out/Environment_Set.Rout'. If \code{FALSE}, the result of checking \code{RNASeqWorkFlowParam} will not be in 'Rscript_out/Environment_Set.Rout'}
}
\value{
None
}
\description{
Process raw reads for RNA-Seq workflow in background.
This function do 5 things :
1. 'Hisat2' : aligns raw reads to reference genome. If \code{indices.optional} in \code{RNASeqWorkFlowParam} is \code{FALSE}, Hisat2 indices will be created.
2. 'Samtools' : converts '.sam' files to '.bam' files.
3. 'Stringtie' : assembles alignments into transcript.
4. 'Gffcompare' : examines how transcripts compare with the reference annotation.
5. 'Stringtie' : creates input files for ballgown, edgeR and DESeq2.
Before running this function, \code{RNASeqEnvironmentSet_CMD()} or\code{RNASeqEnvironmentSet()} must be executed successfully.
If you want to process raw reads for the following RNA-Seq workflow in R shell, please see \code{RNASeqReadProcess()} function.
}
\examples{
\dontrun{
input_file_dir <- system.file(package = "RNASeqWorkflow", "exdata")
exp <- RNASeqWorkFlowParam(path.prefix = "/tmp/", input.path.prefix = input_file_dir, genome.name = "hg19", sample.pattern = "SRR[0-9]",
                           experiment.type = "two.group", main.variable = "treatment", additional.variable = "cell")
## Before run this function, make sure \\code{RNASeqEnvironmentSet_CMD()} (or\\code{RNASeqEnvironmentSet()}) is executed successfully.
RNASeqReadProcess_CMD(RNASeqWorkFlowParam = exp)}
}
\author{
Kuan-Hao Chao
}
