% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RNASeqWorkFlowParam-class}
\alias{RNASeqWorkFlowParam-class}
\alias{RNASeq}
\title{RNASeq
An S4 class for storing RNA-Seq workflow parameters of this package}
\description{
RNASeq
An S4 class for storing RNA-Seq workflow parameters of this package
}
\section{Slots}{

\describe{
\item{\code{os.type}}{'linux' or 'osx'. The operating system type}

\item{\code{python.variable}}{a list with (whether python is valid, python version)}

\item{\code{path.prefix}}{the directory holding installations and analysis results}

\item{\code{input.path.prefix}}{user has to prepared valid 'input_files/' under this directory}

\item{\code{gene.name}}{gene name defined in this RNA-Seq workflow (ex. gene.name.fa, gene.name.gtf)}

\item{\code{sample.pattern}}{sample pattern describing the name of raw fastq.gz files}

\item{\code{experiment.type}}{set the type of the RNA-Seq analysis workflow. Character of one of "two.group", "multi.group.pairs", "multi.group.anova"}

\item{\code{main.variable}}{main sample grouping variable}

\item{\code{additional.variable}}{additional sample information}

\item{\code{indexes.optional}}{logical value whether indexes/ is exit in 'input_files/'}
}}

\examples{
  data(workflowParam)
  class(workflowParam) #"RNASeqWorkFlowParam"
  workflowParam@path.prefix
  workflowParam@input.path.prefix
  workflowParam@gene.name
  workflowParam@sample.pattern
  workflowParam@experiment.type
  workflowParam@main.variable
  workflowParam@additional.variable
}
\author{
Kuan-Hao, Chao
}
