% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RNASeqWorkFlowParam-class}
\alias{RNASeqWorkFlowParam-class}
\alias{RNASeq}
\title{RNASeq workflow
An S4 class for checking and storing RNA-Seq workflow parameters of this package}
\description{
RNASeq workflow
An S4 class for checking and storing RNA-Seq workflow parameters of this package
}
\section{Slots}{

\describe{
\item{\code{os.type}}{'linux' or 'osx'. The operating system type}

\item{\code{python.variable}}{a list storing python environment. (check.answer, python.version)}

\item{\code{path.prefix}}{path prefix of 'gene_data/', 'RNASeq_bin/', 'RNASeq_results/', 'Rscript/' and 'Rscript_out/' directories}

\item{\code{input.path.prefix}}{path prefix of 'input_files/' directory}

\item{\code{genome.name}}{variable of genome name defined in this RNA-Seq workflow (ex. genome.name.fa, genome.name.gtf)}

\item{\code{sample.pattern}}{regular expression of raw fastq.gz files under 'input_files/raw_fastq.gz'}

\item{\code{independent.variable}}{independent variable for the biological experiment design of two-group RNA-Seq workflow}

\item{\code{control.group}}{group name of the control group}

\item{\code{experiment.group}}{group name of the experiment group}

\item{\code{indexes.optional}}{logical value whether 'indexes/' is exit in 'input_files/'}
}}

\examples{
\dontrun{
data(workflowParam)
class(workflowParam) #"RNASeqWorkFlowParam"
workflowParam@path.prefix
workflowParam@input.path.prefix
workflowParam@genome.name
workflowParam@sample.pattern
workflowParam@independent.variable
workflowParam@control.group
workflowParam@experiment.group}
}
\author{
Kuan-Hao, Chao
}
