% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{RNASeqWorkflowParam-constructor}
\alias{RNASeqWorkflowParam-constructor}
\alias{RNASeqWorkflowParam}
\title{Constructor function for RNASeqWorkflowParam objects}
\usage{
RNASeqWorkflowParam(path.prefix = NA, input.path.prefix = NA,
  genome.name = NA, sample.pattern = NA, independent.variable = NA,
  case.group = NA, control.group = NA)
}
\arguments{
\item{path.prefix}{path prefix of 'gene_data/', 'RNASeq_bin/', 'RNASeq_results/', 'Rscript/' and 'Rscript_out/' directories}

\item{input.path.prefix}{path prefix of 'input_files/' directory}

\item{genome.name}{variable of genome name defined in this RNA-Seq workflow (ex. \code{genome.name}.fa, \code{genome.name}.gtf)}

\item{sample.pattern}{Regular expression of paired-end fastq.gz files under 'input_files/raw_fastq.gz'. Expression not includes \code{_[1,2].fastq.gz}.}

\item{independent.variable}{independent variable for the biological experiment design of two-group RNA-Seq workflow}

\item{case.group}{group name of the case group}

\item{control.group}{group name of the control group}
}
\value{
an object of class \code{RNASeqWorkflowParam}
}
\description{
Constructor function for RNASeqWorkflowParam objects
}
\examples{
\dontrun{
exp <- RNASeqWorkflowParam(path.prefix = "/home/RNASeq", input.path.prefix = "/home", genome.name = "hg19", sample.pattern = "SRR[0-9]",
                           independent.variable = "two.group", case.group = "treatment", control.group = "cell")
}
}
\author{
kuan-hao Chao

Kuan-Hao Chao
}
