% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{RNASeqWorkFlowParam-constructor}
\alias{RNASeqWorkFlowParam-constructor}
\alias{RNASeqWorkFlowParam}
\title{Constructor function for RNASeqWorkFlowParam objects}
\usage{
RNASeqWorkFlowParam(path.prefix = NA, input.path.prefix = NA,
  gene.name = NA, sample.pattern = NA, experiment.type = NA,
  main.variable = NA, additional.variable = NA)
}
\arguments{
\item{path.prefix}{the directory holding installations and analysis results}

\item{input.path.prefix}{user has to prepared valid 'input_files/' under this directory}

\item{gene.name}{gene name defined in this RNA-Seq workflow (ex. gene.name.fa, gene.name.gtf)}

\item{sample.pattern}{sample pattern describing the name of raw fastq.gz files}

\item{experiment.type}{set the type of the RNA-Seq analysis workflow. Character of one of "two.group", "pair-wise.group", "multi.group"}

\item{main.variable}{main sample grouping variable}

\item{additional.variable}{additional sample information}
}
\value{
an object of class \code{RNASeqWorkFlowParam}
}
\description{
Constructor function for RNASeqWorkFlowParam objects
}
\author{
kuan-hao Chao
}
