% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/rnaseq_raw_read_process_hisat2_samtools_stringtie_ballgown_related.R
\name{SamtoolsToBam}
\alias{SamtoolsToBam}
\title{use 'samtools' to sort and convert the SAM files to BAM}
\usage{
SamtoolsToBam(path.prefix, genome.name, sample.pattern,
  num.parallel.threads = 8)
}
\description{
use 'samtools' to sort and convert the SAM files to BAM
}
