% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisat2_samtools_stringtie_ballgown_related.R
\name{StringTieMergeTrans}
\alias{StringTieMergeTrans}
\title{stringtie merge transcripts from all samples}
\usage{
StringTieMergeTrans(path.prefix, gene.name, sample.pattern,
  num.parallel.threads = 8)
}
\description{
stringtie merge transcripts from all samples
}
